+++ frontend/pages/admin/IntegrationsPage/cards/Certificates/_styles.scss
@@ -1,28 +1,6 @@
 .certificates-integration {
 
-
   &__page-description {
     margin: 0 0 $pad-xlarge;
-  }
-
-  &__add-cert-card {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: $pad-medium;
-  }
-
-  &__add-cert-card-content {
-    text-align: center;
-
-    p {
-      margin: 0;
-    }
-  }
-
-  &__add-cert-card-title {
-    font-size: $small;
-    font-weight: $bold;
-    margin-bottom: $pad-small;
   }
 }
