+++ frontend/pages/admin/IntegrationsPage/cards/Certificates/components/AddCertAuthorityCard/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./AddCertAuthorityCard";
